Skilt ut NaisProxyCustomizer til egen fil og bønne.
